requirements.txt
flask
flask-cors
pillow
flask_sqlalchemy